pyproject.toml
[tool.poetry]
name = "slackbot"
version = "0.1.0"
description = ""
authors = ["Matheus Deister <matheus.deister@vinta.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
langchain-openai = "^0.1.17"
tiktoken = "^0.7.0"
atlassian-python-api = "^3.41.14"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.9"
sqlite-vss = "^0.1.2"
lxml = "^5.2.2"
markdownify = "^0.13.1"
langchainhub = "^0.1.20"
slack-bolt = "^1.19.1"
slackstyler = "^0.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
black = "^24.4.2"
datamodel-code-generator = "^0.25.8"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
